reviewlist component moved to its own directory
